requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Graph Database
neo4j==5.16.0

# LangChain & GraphRAG
langchain==0.1.4
langchain-community==0.0.16
langchain-openai==0.0.5

# Vector Embeddings
sentence-transformers==2.3.1
openai==1.10.0

# SQL Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
sqlite3-python==1.0.0

# Data Processing
numpy==1.26.3
pandas==2.1.4

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Utilities
python-multipart==0.0.6
